[Discover] Create getRowIndicatorColor extension
