+++ src/plugins/discover/public/context_awareness/types.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import type { CustomCellRenderer } from '@kbn/unified-data-table';
+import type { CustomCellRenderer, UnifiedDataTableProps } from '@kbn/unified-data-table';
 import type { DocViewsRegistry } from '@kbn/unified-doc-viewer';
 import type { DataTableRecord } from '@kbn/discover-utils';
 import type { DataView } from '@kbn/data-views-plugin/common';
@@ -38,4 +38,5 @@ export interface Profile {
   getCellRenderers: () => CustomCellRenderer;
   getDocViewer: (params: DocViewerExtensionParams) => DocViewerExtension;
   getDefaultAppState: (params: DefaultAppStateExtensionParams) => DefaultAppStateExtension;
+  getRowIndicatorColor: UnifiedDataTableProps['getRowIndicatorColor'];
 }
